libgomp: don't use GOMP_USE_ALIGNED_WORK_SHARES on Windows
In case GOMP_USE_ALIGNED_WORK_SHARES is defined, various places
use gomp_aligned_alloc() instead of normal allocations, but then don't
use gomp_aligned_free() in the end, but normal free.

On most platforms this doesn't make a difference, but on Windows
gomp_aligned_alloc() and free() can't be mixed and free() might crash.

Work around this for now by disabling GOMP_USE_ALIGNED_WORK_SHARES
in those cases and add a note on what needs to be done to fix it.